+++ crashtracker-ffi/src/crash_info/stackframe.rs
@@ -4,10 +4,9 @@
 use ::function_name::named;
 use datadog_crashtracker::{BuildIdType, FileType, StackFrame};
 use ddcommon_ffi::{
-    slice::AsBytes, wrap_with_void_ffi_result, CharSlice, Error, Handle, ToInner, VoidResult,
+    slice::AsBytes, utils::ToHexStr, wrap_with_void_ffi_result, CharSlice, Error, Handle, ToInner,
+    VoidResult,
 };
-
-use ddcommon_ffi::ToHexStr;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 //                                              FFI API                                           //
